Do not reset sources if REPO_BRANCH is a tag

+++ scripts/update_repo.sh
@@ -80,7 +80,9 @@ if [ "x${OPENWRT_CUR_DIR}" != "x${OPENWRT_COMPILE_DIR}" ] && [ -d "${OPENWRT_COM
   git -C "${OPENWRT_CUR_DIR}" remote set-url origin "${REPO_URL}"
   git -C "${OPENWRT_CUR_DIR}" fetch
   git -C "${OPENWRT_CUR_DIR}" checkout "${REPO_BRANCH}"
-  git -C "${OPENWRT_CUR_DIR}" reset --hard origin/${REPO_BRANCH}
+  if ! $(git -C "${OPENWRT_CUR_DIR}" tag | grep -q "${REPO_BRANCH}"); then
+    git -C "${OPENWRT_CUR_DIR}" reset --hard origin/${REPO_BRANCH}
+  fi
 else
   git clone -b "${REPO_BRANCH}" "${REPO_URL}" "${OPENWRT_CUR_DIR}"
 fi
